Bugfix flexplorer: Avoid exception when sorting size column.
- Bug came in with cb57e2fa9de25536079e38d05118ff5ea296d5f0
- Filesize has been changed from signed to unsigned type.

+++ src/fpmdich.cpp
@@ -138,7 +138,7 @@ int FlexplorerMdiChild::GetSelectedFilesCount() const
     return selectedFilesCount;
 }
 
-int FlexplorerMdiChild::GetSelectedFilesByteSize() const
+unsigned FlexplorerMdiChild::GetSelectedFilesByteSize() const
 {
     return selectedFilesByteSize;
 }
@@ -713,7 +713,7 @@ void FlexplorerMdiChild::SelectionChanged(
         {
             selectedFilesCount += pmOne;
             auto variant = model->data(index, Qt::DisplayRole);
-            selectedFilesByteSize += pmOne * variant.toInt();
+            selectedFilesByteSize += pmOne * static_cast<int>(variant.toUInt());
         }
     };
